requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
idna==2.10
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pyTelegramBotAPI==3.7.6
PyYAML==5.4.1
requests==2.25.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
